ml-service/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
joblib>=1.3.0